backend/requirements.txt
fastapi
starlette
chainlit
google-generativeai
python-dotenv
starlette
websockets
pymongo
bcrypt
pyjwt
ollama